fix: lower model precision in forward pass

+++ configs.py
@@ -10,7 +10,7 @@ class GPTConfig(eqx.Module):
     n_heads: int = 6
     n_embed: int = 384
     dropout: float = 0.2
-    dtype: jnp.dtype = jnp.bfloat16
+    dtype: jnp.dtype = jnp.float32
 
     @classmethod
     def from_preset(cls, name: str):
